+++ backoffice/src/componenets/Table/Table.js
@@ -1,15 +1,13 @@
 import "./Table.scss"
 import {isEmpty} from 'lodash';
 import {Link} from "react-router-dom";
-export default ({title, headers, rows, navigationButton = {}, actionButton = {}, pager, emptyElement}) => {
+export default ({title, headers, rows, navigationButton = {}, actionButton = {}, pager, emptyElement, limitTableSize = false}) => {
   const {path, text} = navigationButton;
 
   return <section className="table">
 
     <section className="top">
-      <span className="title">
-        {title}
-      </span>
+      <span className="title">{title}</span>
 
       {!isEmpty(navigationButton) && <div className="navigation-button-wrapper">
         <Link to={path} className="navigation-button">{text}</Link>
@@ -20,7 +18,7 @@ export default ({title, headers, rows, navigationButton = {}, actionButton = {},
       </div> }
     </section>
 
-    <div className="table-wrapper">
+    <div className={limitTableSize ? 'table-wrapper' : ''}>
       <table>
         <thead>
         <tr>
@@ -30,7 +28,7 @@ export default ({title, headers, rows, navigationButton = {}, actionButton = {},
 
         <tbody>
         {rows.map((row, key) => <tr key={key}>
-          {row.map((cell, key) => <td key={key} title={`${headers[key]}: ${cell}`}>{cell}</td>)}
+          {row.map((cell, key) => <td key={key} title={headers[key]}>{cell}</td>)}
         </tr>)}
         {isEmpty(rows) && <tr><td className="empty-row" colSpan={5}>
           {emptyElement}
@@ -39,8 +37,6 @@ export default ({title, headers, rows, navigationButton = {}, actionButton = {},
       </table>
     </div>
 
-    {pager && <section className="bottom">
-      {pager}
-    </section>}
+    {pager && <section className="bottom">{pager}</section>}
   </section>
 };
