+++ backoffice_backend/envs/.local.docker.env
@@ -1,6 +1,7 @@
 PORT=80
 dbURL=mongodb://backoffice_backend_db/open_pesion_dashboard
 storageAddress=http://storage
+monthlyAddress=http://monthly
 tempStorageFiles=/home/app/files
 saltRounds=10
 tokenTimeToLive=86400
